+++ R/data_summary.R
@@ -11,9 +11,10 @@
 #' @param remove_na Logical. If `TRUE`, missing values are omitted from the
 #' grouping variable. If `FALSE` (default), missing values are included as a
 #' level in the grouping variable.
-#' @param allow_multiple Logical, if `FALSE` (default), each expression must
-#' return a single value. If `TRUE`, expressions can return more than one
-#' value, which will result in a data frame with more rows.
+#' @param suffix Character vector or a list of character vectors, indicating the
+#' suffixes to be added to the new variable names. This is useful when the
+#' summary function returns more than one value (e.g., `quantile()`). If a list,
+#' it should have the same length as the number of expressions in `...`.
 #' @param ... One or more named expressions that define the new variable name
 #' and the function to compute the summary statistic. Example:
 #' `mean_sepal_width = mean(Sepal.Width)`. The expression can also be provided
@@ -54,7 +55,7 @@
 #'   by = c("am", "gear")
 #' )
 #'
-#' # allow more than one-row-summaries for expressions with allow_multiple=FALSE
+#' # allow more than one-column-summaries for expressions
 #' d <- data.frame(
 #'   x = rnorm(100, 1, 1),
 #'   y = rnorm(100, 2, 2),
@@ -65,16 +66,16 @@
 #'   d,
 #'   quant_x = quantile(x, c(0.25, 0.75)),
 #'   quant_y = quantile(y, c(0.25, 0.75)),
-#'   allow_multiple = TRUE
+#'   suffix = c("Q1", "Q3")
 #' )
 #'
-#' # also works with grouped data
+#' # also works with grouped data, and multiple suffixes
 #' data_summary(
 #'   d,
 #'   quant_x = quantile(x, c(0.25, 0.75)),
-#'   quant_y = quantile(y, c(0.25, 0.75)),
+#'   quant_y = quantile(y, c(0.1, 0.9)),
 #'   by = "groups",
-#'   allow_multiple = TRUE
+#'   suffix = list(c("Q1", "Q3"), c("10perc", "90perc"))
 #' )
 #'
 #' @export
@@ -84,8 +85,20 @@ data_summary <- function(x, ...) {
 
 
 #' @export
-data_summary.matrix <- function(x, ..., by = NULL, remove_na = FALSE) {
-  data_summary(as.data.frame(x), ..., by = by, remove_na = remove_na)
+data_summary.matrix <- function(
+  x,
+  ...,
+  by = NULL,
+  remove_na = FALSE,
+  suffix = NULL
+) {
+  data_summary(
+    as.data.frame(x),
+    ...,
+    by = by,
+    remove_na = remove_na,
+    suffix = suffix
+  )
 }
 
 
@@ -104,7 +117,7 @@ data_summary.data.frame <- function(
   ...,
   by = NULL,
   remove_na = FALSE,
-  allow_multiple = FALSE
+  suffix = NULL
 ) {
   dots <- eval(substitute(alist(...)))
 
@@ -117,10 +130,10 @@ data_summary.data.frame <- function(
 
   if (is.null(by)) {
     # when we have no grouping, just compute a one-row summary
-    summarise <- .process_datasummary_dots(dots, x, allow_multiple)
+    summarise <- .process_datasummary_dots(dots, x, suffix)
     # coerce to data frame
-    out <- list2DF(summarise)
-    colnames(out) <- vapply(summarise, function(cn) names(cn)[1], character(1))
+    out <- as.data.frame(t(summarise))
+    colnames(out) <- names(summarise)
   } else {
     # sanity check - is "by" a character string?
     if (!is.character(by)) {
@@ -156,16 +169,13 @@ data_summary.data.frame <- function(
         return(NULL)
       }
       # summarize data
-      summarise <- .process_datasummary_dots(dots, s, allow_multiple)
+      summarise <- .process_datasummary_dots(dots, s, suffix)
       # coerce to data frame
-      summarised_data <- as.data.frame(summarise, row.names = NULL)
+      summarised_data <- as.data.frame(t(unlist(summarise)))
       # bind grouping-variables and values
       summarised_data <- cbind(s[1, by], summarised_data)
       # make sure we have proper column names
-      colnames(summarised_data) <- c(
-        by,
-        stats::na.omit(unlist(lapply(summarise, names)))
-      )
+      colnames(summarised_data) <- c(by, names(summarise))
       summarised_data
     })
     out <- do.call(rbind, out)
@@ -180,7 +190,13 @@ data_summary.data.frame <- function(
 
 
 #' @export
-data_summary.grouped_df <- function(x, ..., by = NULL, remove_na = FALSE) {
+data_summary.grouped_df <- function(
+  x,
+  ...,
+  by = NULL,
+  remove_na = FALSE,
+  suffix = NULL
+) {
   # extract group variables
   grps <- attr(x, "groups", exact = TRUE)
   group_variables <- data_remove(grps, ".rows")
@@ -191,13 +207,13 @@ data_summary.grouped_df <- function(x, ..., by = NULL, remove_na = FALSE) {
   # remove information specific to grouped df's
   attr(x, "groups") <- NULL
   class(x) <- "data.frame"
-  data_summary(x, ..., by = by, remove_na = remove_na)
+  data_summary(x, ..., by = by, remove_na = remove_na, suffix = suffix)
 }
 
 
 # helper -----------------------------------------------------------------------
 
-.process_datasummary_dots <- function(dots, data, allow_multiple = FALSE) {
+.process_datasummary_dots <- function(dots, data, suffix = NULL) {
   out <- NULL
   if (length(dots)) {
     # we check for character vector of expressions, in which case
@@ -236,39 +252,45 @@ data_summary.grouped_df <- function(x, ..., by = NULL, remove_na = FALSE) {
       }
     }
 
+    # check if we have enough suffixes for the number of expressions
+    if (is.list(suffix) && length(suffix) != length(dots)) {
+      insight::format_error(
+        "If `suffix` is a list of suffixes, it should have the same length as the number of expressions."
+      )
+    }
+
     out <- lapply(seq_along(dots), function(i) {
       new_variable <- .get_new_dots_variable(dots, i, data)
       if (inherits(new_variable, c("bayestestR_ci", "bayestestR_eti"))) {
         stats::setNames(new_variable, c("CI", "CI_low", "CI_high"))
       } else {
-        stats::setNames(new_variable, names(dots)[i])
+        # if we have a list of suffixes, we just want the current one that
+        # is related to the current expression
+        if (is.list(suffix)) {
+          current_suffix <- suffix[[i]]
+        } else {
+          current_suffix <- suffix
+        }
+        # if we don't have suffixes for multiple columns, but expression
+        # returns multiple columns, we get NA column names - we warn the user
+        if (is.null(current_suffix) && length(new_variable) > 1) {
+          insight::format_error(
+            "Each expression must return a single value, or you must provide a character vector of suffixes for the new variable names using `suffix`."
+          )
+        } else if (
+          !is.null(current_suffix) &&
+            length(current_suffix) != length(new_variable)
+        ) {
+          insight::format_error(
+            "All expressions must return the same number of values and argument `suffix` must have the same length as the result of the summary expression(s)."
+          )
+        }
+        stats::setNames(new_variable, paste0(names(dots)[i], current_suffix))
       }
     })
   }
 
-  # check for correct length of output - must be a single value!
-  if (!isTRUE(allow_multiple)) {
-    # Exception: bayestestR::ci()
-    wrong_length <- !inherits(out, "bayestestR_ci") &&
-      !inherits(out, "bayestestR_eti") &&
-      lengths(out) != 1
-
-    if (any(wrong_length)) {
-      insight::format_error(
-        paste0(
-          "Each expression must return a single value. Following expression(s) returned more than one value: ",
-          text_concatenate(
-            vapply(dots[wrong_length], insight::safe_deparse, character(1)),
-            enclose = "\""
-          ),
-          "."
-        ),
-        "\nUse `allow_multiple = TRUE` to allow expressions to return more than one value."
-      )
-    }
-  }
-
-  out
+  unlist(out)
 }